v2.0b3
* Licence has changed from CC BY-NC-ND to CC BY-ND
* improvement: convert cPassMan to UTF-8 charset
* improvement: all data exchanges concerning items are encrypted based on AES-256
* improvement: new design
* improvement: cpassman uses jquery 1.4.4 and jqueryui 1.8.9
* improvement: drag and drop items into folder (remain some dropable area troubles)
* improvement: allow role access on item base [3160946]
* improvement: new strategy for user password recovery
* improvement: automatic creation of role and folder based on email domain
* improvement: when deleting a user, personal folder and items are deleted
* improvement: LDAP login (thanks to Paul)
* improvement: setting "allow modification for this role" in roles matrix
* improvement: when importing items, 2 options permits to set restriction arguments
* improvement: "copy to clipboard" items are deleted for memory optimization [3177788]
* improvement: dynamic load of items list (to optimze time loading)
* improvement: on cpassman exit, user is fully disconnected
* improvement: timezone is customizable by admin [19]
* improvement: LDAP extension is checked before activating this functionality [25]
* improvement: manage folders page has been reviewed [30]
* improvement: By default 'admin' account doesn't use LDAP [41]
* improvement: If LDAP, not existing user in CPM is automatically created [43]
* improvement: duplicate item from find page
* improvement: add session_key verification for POST actions
* fixed: problem with slash in label [3141743]
* fixed: password in cleartext in log [3141167]
* fixed: deleted pws show up in the search [3160582]
* fixed: GUI corrections [3160611]
* fixed: delete session of a deleted user [16]
* fixed: charset problem on install under IE [17]
* fixed: JSON error on search page [18]
* fixed: PHP version is checked during upgrade and install [19]
* fixed: if no roles and folders then some buttons are disabled [20]
* fixed: [23 - 29 - 32 - 33 - 36 - 37 - 44 - 47 - 53 - 56 - 58 - 48 - 65 - 68 - 63 - 69]

+++ sources/users.queries.php
@@ -89,6 +89,12 @@ if ( !empty($_POST['type']) ){
 
         ## ADD NEW USER ##
         case "add_new_user":
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
             // Check if user already exists
             $db->query("SELECT id, fonction_id, groupes_interdits, groupes_visibles FROM ".$pre."users WHERE login LIKE '".mysql_real_escape_string(stripslashes($_POST['login']))."'");
             $data = $db->fetch_array();
@@ -199,6 +205,12 @@ SendEmail(
 
         ## DELETE USER ##
         case "supprimer_user":
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
         	//delete user in database
             $db->query("DELETE FROM ".$pre."users WHERE id = ".$_POST['id']);
 
@@ -234,70 +246,106 @@ SendEmail(
 
         ## UPDATE PASSWORD OF USER ##
         case "modif_mdp_user":
-           $db->query_update(
-                "users",
-                array(
-                    'pw' => encrypt(string_utf8_decode($_POST['newmdp']))
-                ),
-                "id = ".$_POST['id']
-            );
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
 
-            echo '$("#div_loading").hide()';  //hide loading div
+			$db->query_update(
+			     "users",
+			     array(
+			         'pw' => encrypt(string_utf8_decode($_POST['newmdp']))
+			     ),
+			     "id = ".$_POST['id']
+			 );
+
+			 echo '$("#div_loading").hide()';  //hide loading div
         break;
 
         ## UPDATE EMAIL OF USER ##
         case "modif_mail_user":
-           $db->query_update(
-                "users",
-                array(
-                    'email' => $_POST['newemail']
-                ),
-                "id = ".$_POST['id']
-            );
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
+			$db->query_update(
+			     "users",
+			     array(
+			         'email' => $_POST['newemail']
+			     ),
+			     "id = ".$_POST['id']
+			 );
         break;
 
         // UPDATE CAN CREATE ROOT FOLDER RIGHT
         case "can_create_root_folder":
-           $db->query_update(
-                "users",
-                array(
-                    'can_create_root_folder' => $_POST['value']
-                ),
-                "id = ".$_POST['id']
-            );
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
+			$db->query_update(
+			     "users",
+			     array(
+			         'can_create_root_folder' => $_POST['value']
+			     ),
+			     "id = ".$_POST['id']
+			 );
         break;
 
         ## UPDATE MANAGER RIGHTS FOR USER ##
         case "gestionnaire":
-           $db->query_update(
-                "users",
-                array(
-                    'gestionnaire' => $_POST['value']
-                ),
-                "id = ".$_POST['id']
-            );
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
+			$db->query_update(
+			     "users",
+			     array(
+			         'gestionnaire' => $_POST['value']
+			     ),
+			     "id = ".$_POST['id']
+			 );
         break;
 
         ## UPDATE ADMIN RIGHTS FOR USER ##
         case "admin":
-           $db->query_update(
-                "users",
-                array(
-                    'admin' => $_POST['value']
-                ),
-                "id = ".$_POST['id']
-            );
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
+			$db->query_update(
+			     "users",
+			     array(
+			         'admin' => $_POST['value']
+			     ),
+			     "id = ".$_POST['id']
+			 );
         break;
 
         ## UPDATE PERSONNAL FOLDER FOR USER ##
         case "personal_folder":
-           $db->query_update(
-                "users",
-                array(
-                    'personal_folder' => $_POST['value']
-                ),
-                "id = ".$_POST['id']
-            );
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
+			$db->query_update(
+			     "users",
+			     array(
+			         'personal_folder' => $_POST['value']
+			     ),
+			     "id = ".$_POST['id']
+			 );
         break;
 
         //CHANGE USER FUNCTIONS
@@ -324,6 +372,12 @@ SendEmail(
         break;
 
         case "change_user_functions";
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
             //save data
             $db->query_update(
                 "users",
@@ -377,6 +431,12 @@ SendEmail(
         break;
 
         case "change_user_autgroups";
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
             //save data
             $db->query_update(
                 "users",
@@ -403,6 +463,12 @@ SendEmail(
 
         //CHANGE FORBIDDEN GROUPS
         case "open_div_forgroups";
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
             $text = "";
             //Refresh list of existing functions
             $data_user = $db->fetch_row("SELECT groupes_interdits FROM ".$pre."users WHERE id = ".$_POST['id']);
@@ -457,6 +523,12 @@ SendEmail(
 
         ## UNLOCK USER ##
         case "unlock_account":
+        	//Check KEY
+        	if ($_POST['key'] != $_SESSION['key']) {
+        		//error
+        		exit();
+        	}
+
             $db->query_update(
                 "users",
                 array(
